CCSDSSpacePacketPrimaryHeader.hh - getAPIDAsInteger() now returns uint16_t instead of size_t

+++ includes/CCSDSSpacePacketPrimaryHeader.hh
@@ -159,7 +159,7 @@ public:
 
 public:
 	/** Returns APDI as an integer. */
-	inline size_t getAPIDAsInteger() const {
+	inline uint16_t getAPIDAsInteger() const {
 		return apid.to_ulong();
 	}
